+++ app/jobs/create_stats_job.rb
@@ -1,6 +1,6 @@
 class CreateStatsJob < ApplicationJob
   def perform
-    department_ids = Department.all.collect(&:id).push(nil)
+    department_ids = Department.pluck(:id).push(nil)
     department_ids.each do |department_id|
       Stats::CreateStat.call(department_id: department_id)
     end
